+++ clients/fides-js/src/lib/initialize.ts
@@ -369,7 +369,7 @@ export const initialize = async ({
   // Call extensions
   // DEFER(PROD#1439): This is likely too late for the GPP stub.
   // We should move stub code out to the base package and call it right away instead.
-  await setupExtensions(options);
+  await setupExtensions({ options, experience: effectiveExperience });
 
   // return an object with the updated Fides values
   return {